+++ tests/src/test/scala/org/apache/openwhisk/core/database/ArtifactActivationStoreBehaviorBase.scala
@@ -15,9 +15,19 @@
  * limitations under the License.
  */
 
-package whisk.core.database.test.behavior
+package org.apache.openwhisk.core.database
 
-trait ActivationStoreBehavior
-    extends ActivationStoreBehaviorBase
-    with ActivationStoreCRUDBehaviors
-    with ActivationStoreQueryBehaviors
+import akka.stream.ActorMaterializer
+import org.apache.openwhisk.core.controller.test.WhiskAuthHelpers
+import org.apache.openwhisk.core.database.test.behavior.ActivationStoreBehaviorBase
+import org.scalatest.FlatSpec
+
+trait ArtifactActivationStoreBehaviorBase extends FlatSpec with ActivationStoreBehaviorBase {
+  override def storeType = "Artifact"
+
+  override val context = UserContext(WhiskAuthHelpers.newIdentity())
+
+  override lazy val activationStore = {
+    ArtifactActivationStoreProvider.instance(actorSystem, ActorMaterializer.create(actorSystem), logging)
+  }
+}
